requirements.txt
numpy
pandas
matplotlib
seaborn
kafka-python
jupyter
mysql-connector-python
-e .